long nonce

+++ veriblock-core/src/main/java/org/veriblock/core/contracts/FilteredBlock.java
@@ -24,7 +24,7 @@ public interface FilteredBlock {
 
     int getDifficulty();
 
-    int getNonce();
+    long getNonce();
 
     int getTotalRegularTransactions();
